refactor(systemevent): add domain Repository interface, move params to domain package
